refactor: merge observability/ into monitoring/
Fold the 2-file observability package (tracing.py) into monitoring/ alongside
metrics/health/alerts, removing a top-level package and de-overloading the
'observability' name (vs the unrelated ObservabilityContainer). Repoint 5
importers and move the tracing tests to mirror the source. Drop the redundant
observability import-boundary rule (the monitoring rule already covers it) and
update architecture docs.

Verified: 976 targeted tests, import-boundary check, link audit, agent-regenerate.
