--- tests/unit/gpt_trader/observability/__init__.py
@@ -1,1 +1,0 @@
-"""Observability tests."""
